added frontend to DarkSUSY_generic_wimp_6.2.2 + updated WIMP standalone; bugfix in MSSM standalone

+++ Backends/src/frontends/DarkSUSY_MSSM_6_2_2.cpp
@@ -520,10 +520,10 @@ BE_NAMESPACE
     std::vector<double> couplings;
     couplings.clear();
     //TB FIXME: correct type conversion!?
-    couplings.push_back(gg(1,1).re); // gps
-    couplings.push_back(gg(1,2).re); // gns
-    couplings.push_back(gg(4,1).re); // gpa
-    couplings.push_back(gg(4,2).re); // gna
+    couplings.push_back((double)gg(1,1).re); // gps
+    couplings.push_back((double)gg(1,2).re); // gns
+    couplings.push_back((double)gg(4,1).re); // gpa
+    couplings.push_back((double)gg(4,2).re); // gna
     
     
     return couplings;
